fix(icons): prefer "file" category in `mock_nvim_web_devicons()`
This leads to a slightly different logic if both `name` and `ext` are
supplied to `require('nvim-web-devicons').get_icon()` (like currently is
done in 'nvim-telescope/telescope.nvim', for example).

In practice this seems like a better choice because it accounts for
special file names at the cost of not preferring `ext` if it conflicts
with `name` (which rarely happens) and very small overhead of
recomputing extension (which assumed to already be computed by the
caller).