feat(icons): add many new filetype icons
All new filetype icons are explicitly supported by vanilla Neovim in
a sense that they can be the output of `vim.filetype.match()`. Previous
set was computed based on `vim.fn.getcompletion('', 'filetype')`, which
returns filetypes that have a dedicated 'ftplugin', 'syntax', or
'indent' file. This turned out to be not enough, as there were several
PRs adding missing filetypes.