fix(icons)!: make `config.use_file_extension` be called only once
Previously it was called for every suffix after the dot, but it was not
consistent:
- Cases like 'file.not.recognized.extension' call for every suffix.
- Cases like 'file.is.recognized.lua' call only first time because
  `MiniIcons.get('extension', 'is.recognized.lua')` returned not
  fallback icon right away. This made wrong impression that making
  `use_file_extension()` return `false` for 'recognized.lua' would block
  this case, yet it did not.

The solution is to make `use_file_extension()` be called only once and
prompt users to take care of the whole situation themselves.

Multiple calls also were done because there were no explicit resolution
in 'extension' category. Now there is.